some changes to prevent integer size conversions attempt to fix weird score bugs
git-svn-id: https://svn.code.sf.net/p/ultrastardx/svn/trunk@1932 b956fd51-792f-4845-bead-9b4dfca2ff2c

+++ src/screens/UScreenSing.pas
@@ -859,7 +859,7 @@ end;
 
 procedure TScreenSing.OnSentenceEnd(SentenceIndex: cardinal);
 var
-  PlayerIndex: integer;
+  PlayerIndex: byte;
   CurrentPlayer: PPLayer;
   CurrentScore: real;
   Line:      PLine;
